Refactored Main Header it's own Component Removed the extra header from Volunteers Added a blank Dashboard to navigation
