+++ src/caretogether-pwa/src/App.tsx
@@ -1,16 +1,11 @@
 import React from 'react';
 import clsx from 'clsx';
 import { fade, makeStyles } from '@material-ui/core/styles';
-import { Typography, InputBase, CssBaseline, AppBar, Toolbar, IconButton, Badge, Drawer, Divider, List, BottomNavigation, BottomNavigationAction, useMediaQuery, useTheme } from '@material-ui/core';
+import { Typography, CssBaseline, IconButton, Drawer, Divider, List, BottomNavigation, BottomNavigationAction, useMediaQuery, useTheme } from '@material-ui/core';
 import PermPhoneMsgIcon from '@material-ui/icons/PermPhoneMsg';
-import EmojiPeopleIcon from '@material-ui/icons/EmojiPeople';
 import PeopleIcon from '@material-ui/icons/People';
-import SearchIcon from '@material-ui/icons/Search';
 import ChevronLeftIcon from '@material-ui/icons/ChevronLeft';
-import MenuIcon from '@material-ui/icons/Menu';
-import NotificationsIcon from '@material-ui/icons/Notifications';
-import AssignmentIcon from '@material-ui/icons/Assignment';
-import AssignmentTurnedInIcon from '@material-ui/icons/AssignmentTurnedIn';
+import DashboardIcon from '@material-ui/icons/Dashboard';
 import { Route, Switch, Redirect, BrowserRouter as Router, Link } from "react-router-dom";
 import { ListItemLink } from './Components/ListItemLink';
 import { Arrangements } from './Components/Arrangements';
@@ -23,6 +18,8 @@ import { Communities } from './Components/Communities';
 import { useRecoilValue } from 'recoil';
 import { locationNameData, organizationNameData } from './Model/ConfigurationModel';
 import { Volunteers } from './Components/Volunteers';
+import Header from './Components/Header';
+import { Dashboard } from './Components/Dashboard';
 
 const copyrightStyles = makeStyles((theme) => ({
   copyright: {
@@ -48,9 +45,6 @@ const useStyles = makeStyles((theme) => ({
   root: {
     display: 'flex',
   },
-  toolbar: {
-    paddingRight: 24, // keep right padding when drawer closed
-  },
   drawerHeader: {
     display: 'flex',
     alignItems: 'center',
@@ -66,70 +60,6 @@ const useStyles = makeStyles((theme) => ({
     margin: '0',
     paddingLeft: '8px',
     fontSize: '14px'
-  },
-  appBar: {
-    zIndex: theme.zIndex.drawer + 1,
-    transition: theme.transitions.create(['width', 'margin'], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen,
-    }),
-  },
-  appBarShift: {
-    marginLeft: drawerWidth,
-    width: `calc(100% - ${drawerWidth}px)`,
-    transition: theme.transitions.create(['width', 'margin'], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-  },
-  search: {
-    position: 'relative',
-    borderRadius: theme.shape.borderRadius,
-    backgroundColor: fade(theme.palette.common.white, 0.15),
-    '&:hover': {
-      backgroundColor: fade(theme.palette.common.white, 0.25),
-    },
-    marginLeft: 0,
-    marginRight: theme.spacing(1),
-    width: '100%',
-    [theme.breakpoints.up('sm')]: {
-      marginLeft: theme.spacing(1),
-      width: 'auto',
-    },
-  },
-  searchIcon: {
-    padding: theme.spacing(0, 2),
-    height: '100%',
-    position: 'absolute',
-    pointerEvents: 'none',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  inputRoot: {
-    color: 'inherit',
-  },
-  inputInput: {
-    padding: theme.spacing(1, 1, 1, 0),
-    // vertical padding + font size from searchIcon
-    paddingLeft: `calc(1em + ${theme.spacing(4)}px)`,
-    transition: theme.transitions.create('width'),
-    width: '100%',
-    [theme.breakpoints.up('sm')]: {
-      width: '12ch',
-      '&:focus': {
-        width: '20ch',
-      },
-    },
-  },
-  menuButton: {
-    marginRight: 36,
-  },
-  menuButtonHidden: {
-    display: 'none',
-  },
-  title: {
-    flexGrow: 1,
   },
   drawerPaper: {
     position: 'relative',
@@ -172,6 +102,7 @@ const useStyles = makeStyles((theme) => ({
 
 const mainListItems = (
   <List aria-label="main navigation">
+    <ListItemLink to="/dashboard" primary="Dashboard" icon={<DashboardIcon />} />
     <ListItemLink to="/referrals" primary="Referrals" icon={<PermPhoneMsgIcon />} />
   </List>
 );
@@ -203,40 +134,7 @@ function App() {
     <div className={classes.root}>
       <CssBaseline />
       <Router>
-        <AppBar position="absolute" className={clsx(classes.appBar, (open && !isMobile) && classes.appBarShift)}>
-          <Toolbar className={classes.toolbar} variant="dense">
-            <IconButton
-              edge="start"
-              color="inherit"
-              aria-label="open drawer"
-              onClick={handleDrawerOpen}
-              className={clsx(classes.menuButton, open && classes.menuButtonHidden)}
-            >
-              <MenuIcon />
-            </IconButton>
-            <Typography component="h1" variant="h6" color="inherit" noWrap className={classes.title}>
-              Dashboard
-            </Typography>
-            <div className={classes.search}>
-              <div className={classes.searchIcon}>
-                <SearchIcon />
-              </div>
-              <InputBase
-                placeholder="Search…"
-                classes={{
-                  root: classes.inputRoot,
-                  input: classes.inputInput,
-                }}
-                inputProps={{ 'aria-label': 'search' }}
-              />
-            </div>
-            <IconButton color="inherit">
-              <Badge badgeContent={4} color="secondary">
-                <NotificationsIcon />
-              </Badge>
-            </IconButton>
-          </Toolbar>
-        </AppBar>
+        <Header open={open} handleDrawerOpen={handleDrawerOpen} />
         {isMobile ? null :
           <Drawer
             variant="permanent"
@@ -266,6 +164,9 @@ function App() {
           {/* <Container maxWidth="lg" className={classes.container}> */}
           <React.Suspense fallback={<div>Loading...</div>}>
             <Switch>
+              <Route path="/dashboard">
+                <Dashboard />
+              </Route>
               <Route path="/arrangements">
                 <Arrangements />
               </Route>
@@ -296,20 +197,19 @@ function App() {
             </Switch>
           </React.Suspense>
           {/* </Container> */}
-          {isMobile ? <BottomNavigation
-            value={selectedBottomNavAction}
-            onChange={(_, newValue) => {
-              setSelectedBottomNavAction(newValue);
-            }}
-            showLabels
-            className={classes.stickToBottom}
-          >
-            <BottomNavigationAction component={Link} to="/volunteers" label="Volunteers" icon={<PeopleIcon />} />
-            <BottomNavigationAction component={Link} to="/volunteerApproval" label="Approvals" icon={<EmojiPeopleIcon />} />
-            <BottomNavigationAction component={Link} to="/volunteerApplications" label="Applications" icon={<AssignmentIcon />} />
-            <BottomNavigationAction component={Link} to="/volunteerProgress" label="Progress" icon={<AssignmentTurnedInIcon />} />
-          </BottomNavigation> : null}
         </main>
+        {isMobile ? <BottomNavigation
+          value={selectedBottomNavAction}
+          onChange={(_, newValue) => {
+            setSelectedBottomNavAction(newValue);
+          }}
+          showLabels
+          className={classes.stickToBottom}
+        >
+          <BottomNavigationAction component={Link} to="/dashboard" label="Dashboard" icon={<DashboardIcon />} />
+          <BottomNavigationAction component={Link} to="/referrals" label="Referrals" icon={<PermPhoneMsgIcon />} />
+          <BottomNavigationAction component={Link} to="/volunteers" label="Volunteers" icon={<PeopleIcon />} />
+        </BottomNavigation> : null}
       </Router>
     </div>
   );
